# api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/meetings": {
      "get": {
        "description": "All meetings with state, client, duration, attendees, transcript flag and linked proposal id.",
        "operationId": "MeetingsController_list",
        "parameters": [
          {
            "name": "state",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "DETECTED",
                "AWAITING_TRANSCRIPT",
                "TRANSCRIPT_ATTACHED",
                "CLIENT_MATCHED",
                "UNKNOWN_ATTENDEE",
                "INVOICE_PROPOSED",
                "SENT",
                "SKIPPED"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MeetingListItemDto"
                  }
                }
              }
            }
          }
        },
        "summary": "List meetings",
        "tags": [
          "meetings"
        ]
      }
    },
    "/meetings/{id}": {
      "get": {
        "description": "Full meeting with parsed transcript, match proposals, linked proposal and the evidence/decision chain.",
        "operationId": "MeetingsController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Meeting id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingDetailDto"
                }
              }
            }
          }
        },
        "summary": "Meeting detail",
        "tags": [
          "meetings"
        ]
      }
    },
...
    "/meetings/import-ics": {
      "post": {
        "description": "Fallback ingest path: parse raw .ics text, upsert meetings and run the Loop-1 step.",
        "operationId": "MeetingsController_importIcs",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportIcsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncResultDto"
                }
              }
            }
          }
        },
        "summary": "Import an .ics calendar",
        "tags": [
          "meetings"
        ]
      }
    },
    "/meetings/{id}/transcript": {
      "post": {
        "description": "Saves and parses the transcript, resolves the PROVIDE_TRANSCRIPT task, then runs the Loop-1 continuation (match, price, policy, auto-send or raise REVIEW_INVOICE).",
        "operationId": "MeetingsController_attachTranscript",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Meeting id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttachTranscriptDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingActionResultDto"
                }
              }
            }
          }
        },
        "summary": "Attach a transcript",
        "tags": [
          "meetings"
        ]
      }
    },
    "/meetings/{id}/confirm-client": {
      "post": {
        "description": "Sets the client, resolves the CONFIRM_CLIENT_MATCH task and re-runs the proposal-building continuation.",
        "operationId": "MeetingsController_confirmClient",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Meeting id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmClientDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingActionResultDto"
                }
              }
            }
          }
        },
        "summary": "Confirm the client for a meeting",
        "tags": [
          "meetings"
        ]
      }
    },
    "/meetings/{id}/skip": {
      "post": {
        "description": "Marks the meeting SKIPPED (not billable) and resolves any open transcript/match task.",
        "operationId": "MeetingsController_skip",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Meeting id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SkipMeetingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingDetailDto"
                }
              }
            }
          }
        },
        "summary": "Skip a meeting",
        "tags": [
          "meetings"
        ]
      }
    },
...
    "/detections": {
      "get": {
        "description": "All detections with type, matched client, recoverable value, evidence and state. Filter with ?state= and/or ?type=.",
        "operationId": "DetectionsController_list",
        "parameters": [
          {
            "name": "state",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "OPEN",
                "PROPOSED",
                "RESOLVED",
                "DISMISSED"
              ],
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "QUOTE_NOT_INVOICED",
                "RETAINER_STOPPED",
                "EXPENSE_NOT_RECHARGED"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DetectionDto"
                  }
                }
              }
            }
          }
        },
        "summary": "List ledger detections",
        "tags": [
          "detections"
        ]
      }
    },
    "/detections/{id}/propose": {
      "post": {
        "description": "Builds a single-line InvoiceProposal (provenance LEDGER) from the detection, sets it IN_REVIEW with 20% UK VAT, flips the detection to PROPOSED and raises a REVIEW_INVOICE task. Nothing is written to Xero here — approval lives in ProposalsModule. Idempotent: re-proposing returns the existing proposal.",
        "operationId": "DetectionsController_propose",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetectionProposeResultDto"
                }
              }
            }
          }
        },
        "summary": "Turn a detection into an invoice proposal",
        "tags": [
          "detections"
        ]
      }
    },
    "/detections/{id}/dismiss": {
      "post": {
        "description": "Sets the detection DISMISSED and records the reason (optional) on the audit trail.",
        "operationId": "DetectionsController_dismiss",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DismissDetectionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetectionActionResultDto"
                }
              }
            }
          }
        },
        "summary": "Dismiss a detection",
        "tags": [
          "detections"
        ]
      }
    },
...
  },
  "info": {
    "title": "Robyn API",
    "description": "Robyn turns a freelancer's calendar, contracts and transcripts into invoices in Xero. Every agent behaviour is a state transition that writes to Xero or raises a Task. The LLM parses and proposes at the edges; it never decides and never sends.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "MeetingState": {
        "type": "string",
        "enum": [
          "DETECTED",
          "AWAITING_TRANSCRIPT",
          "TRANSCRIPT_ATTACHED",
          "CLIENT_MATCHED",
          "UNKNOWN_ATTENDEE",
          "INVOICE_PROPOSED",
          "SENT",
          "SKIPPED"
        ]
      },
      "AttendeeDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "jane@fenwick.co.uk"
          },
          "name": {
            "type": "string",
            "example": "Jane Fenwick"
          },
          "organizer": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "email"
        ]
      },
      "MeetingListItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "gcalEventId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "start": {
            "type": "string",
            "description": "ISO timestamp."
          },
          "end": {
            "type": "string",
            "description": "ISO timestamp."
          },
          "durationHours": {
            "type": "number",
            "example": 1.5
          },
          "state": {
            "allOf": [
              {
                "$ref": "#/components/schemas/MeetingState"
              }
            ]
          },
          "clientId": {
            "type": "object",
            "nullable": true
          },
          "clientName": {
            "type": "object",
            "nullable": true
          },
          "attendees": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AttendeeDto"
            }
          },
          "isPersonal": {
            "type": "boolean",
            "description": "No external attendees — auto-skipped."
          },
          "hasTranscript": {
            "type": "boolean"
          },
          "proposalId": {
            "type": "object",
            "nullable": true,
            "description": "Linked InvoiceProposal id, if any."
          },
          "source": {
            "type": "string",
            "example": "ics",
            "description": "'google' | 'ics'"
          }
        },
        "required": [
          "id",
          "gcalEventId",
          "title",
          "start",
          "end",
          "durationHours",
          "state",
          "attendees",
          "isPersonal",
          "hasTranscript",
          "source"
        ]
      },
      "TranscriptSource": {
        "type": "string",
        "enum": [
          "PASTED",
          "GRANOLA",
          "UPLOAD"
        ]
      },
...
      "TranscriptParsedDto": {
        "type": "object",
        "properties": {
          "summary": {
            "type": "object",
            "nullable": true
          },
          "action_points": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "scope_items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScopeItemDto"
            }
          }
        },
        "required": [
          "action_points",
          "scope_items"
        ]
      },
      "TranscriptDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "source": {
            "allOf": [
              {
                "$ref": "#/components/schemas/TranscriptSource"
              }
            ]
          },
          "rawText": {
            "type": "string"
          },
          "parsed": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/TranscriptParsedDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "source",
          "rawText"
        ]
      },
      "MatchProposalDto": {
        "type": "object",
        "properties": {
          "client_id": {
            "type": "string"
          },
          "client_name": {
            "type": "string"
          },
          "confidence": {
            "type": "number",
            "example": 0.82,
            "description": "0..1 confidence — a proposal, never a decision."
          },
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "client_id",
          "client_name",
          "confidence",
          "reason"
        ]
      },
...
      "MeetingProposalDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "state": {
            "allOf": [
              {
                "$ref": "#/components/schemas/InvoiceProposalState"
              }
            ]
          },
          "reference": {
            "type": "object",
            "example": "ROBYN-1a2b3c4d"
          },
          "currency": {
            "type": "string",
            "example": "GBP"
          },
          "subtotal": {
            "type": "number",
            "example": 225
          },
          "taxTotal": {
            "type": "number",
            "example": 45
          },
          "total": {
            "type": "number",
            "example": 270
          },
          "autoSent": {
            "type": "boolean",
            "description": "True if Robyn auto-sent it (autonomy ON)."
          },
          "xeroInvoiceId": {
            "type": "object",
            "nullable": true
          },
          "xeroInvoiceNumber": {
            "type": "object",
            "nullable": true
          },
          "xeroDeepLink": {
            "type": "object",
            "nullable": true
          },
          "policyResult": {
            "$ref": "#/components/schemas/PolicyResultDto"
          },
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProposalLineDto"
            }
          }
        },
        "required": [
          "id",
          "state",
          "currency",
          "subtotal",
          "taxTotal",
          "total",
          "autoSent",
          "policyResult",
          "lines"
        ]
      },
...
      "MeetingDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "gcalEventId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "start": {
            "type": "string",
            "description": "ISO timestamp."
          },
          "end": {
            "type": "string",
            "description": "ISO timestamp."
          },
          "durationHours": {
            "type": "number",
            "example": 1.5
          },
          "state": {
            "allOf": [
              {
                "$ref": "#/components/schemas/MeetingState"
              }
            ]
          },
          "clientId": {
            "type": "object",
            "nullable": true
          },
          "clientName": {
            "type": "object",
            "nullable": true
          },
          "attendees": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AttendeeDto"
            }
          },
          "isPersonal": {
            "type": "boolean",
            "description": "No external attendees — auto-skipped."
          },
          "hasTranscript": {
            "type": "boolean"
          },
          "proposalId": {
            "type": "object",
            "nullable": true,
            "description": "Linked InvoiceProposal id, if any."
          },
          "source": {
            "type": "string",
            "example": "ics",
            "description": "'google' | 'ics'"
          },
          "skipReason": {
            "type": "object",
            "nullable": true,
            "description": "Reason the meeting was skipped."
          },
          "transcript": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/TranscriptDto"
              }
            ]
          },
          "matchProposals": {
            "nullable": true,
            "description": "LLM fuzzy match proposals (only when unknown/ambiguous).",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MatchProposalDto"
            }
          },
          "proposal": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/MeetingProposalDto"
              }
            ]
          },
          "auditTrail": {
            "description": "The evidence/decision chain, newest first.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuditEventDto"
            }
          }
        },
        "required": [
          "id",
          "gcalEventId",
          "title",
          "start",
          "end",
          "durationHours",
          "state",
          "attendees",
          "isPersonal",
          "hasTranscript",
          "source",
          "auditTrail"
        ]
      },
      "ConnectionStatus": {
        "type": "string",
        "enum": [
          "LIVE",
          "FALLBACK",
          "DOWN"
        ]
      },
      "SyncResultDto": {
        "type": "object",
        "properties": {
          "source": {
            "type": "string",
            "example": "ics",
            "description": "'google' | 'ics'"
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ConnectionStatus"
              }
            ]
          },
          "label": {
            "type": "object",
            "nullable": true
          },
          "detail": {
            "type": "object",
            "nullable": true
          },
          "imported": {
            "type": "number",
            "description": "New meetings created this run."
          },
          "updated": {
            "type": "number",
            "description": "Existing meetings refreshed this run."
          },
          "processed": {
            "type": "number",
            "description": "Meetings the Loop-1 step ran on this run."
          },
          "lastSyncAt": {
            "type": "string",
            "description": "ISO timestamp of this sync."
          }
        },
        "required": [
          "source",
          "status",
          "imported",
          "updated",
          "processed",
          "lastSyncAt"
        ]
      },
      "ImportIcsDto": {
        "type": "object",
        "properties": {
          "icsText": {
            "type": "string",
            "description": "Raw .ics calendar text to ingest."
          }
        },
        "required": [
          "icsText"
        ]
      },
      "AttachTranscriptDto": {
        "type": "object",
        "properties": {
          "rawText": {
            "type": "string",
            "description": "The pasted / uploaded transcript text."
          },
          "source": {
            "default": "PASTED",
            "allOf": [
              {
                "$ref": "#/components/schemas/TranscriptSource"
              }
            ]
          }
        },
        "required": [
          "rawText"
        ]
      },
      "MeetingActionResultDto": {
        "type": "object",
        "properties": {
          "meeting": {
            "$ref": "#/components/schemas/MeetingDetailDto"
          },
          "proposal": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/MeetingProposalDto"
              }
            ]
          },
          "xeroError": {
            "type": "object",
            "nullable": true,
            "description": "Populated when an attempted Xero write failed (state kept locally)."
          }
        },
        "required": [
          "meeting"
        ]
      },
      "ConfirmClientDto": {
        "type": "object",
        "properties": {
          "clientId": {
            "type": "string",
            "description": "The Client to bill this meeting to."
          }
        },
        "required": [
          "clientId"
        ]
      },
      "SkipMeetingDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "description": "Why the meeting is not billable."
          }
        }
      },
...
      "DetectionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "QUOTE_NOT_INVOICED",
              "RETAINER_STOPPED",
              "EXPENSE_NOT_RECHARGED"
            ]
          },
          "clientId": {
            "type": "object",
            "nullable": true,
            "description": "Matched client id, if resolved"
          },
          "clientName": {
            "type": "string",
            "description": "Resolved client display name"
          },
          "valueGbp": {
            "type": "number",
            "example": 4200,
            "description": "Recoverable value in GBP"
          },
          "summary": {
            "type": "string",
            "description": "Human explanation of the leak (from evidence.summary)"
          },
          "source": {
            "type": "string",
            "enum": [
              "xero",
              "local"
            ],
            "description": "Which data path produced this detection"
          },
          "evidence": {
            "type": "object",
            "additionalProperties": true,
            "description": "Flexible per-detector evidence blob (always includes summary + source)"
          },
          "state": {
            "type": "string",
            "enum": [
              "OPEN",
              "PROPOSED",
              "RESOLVED",
              "DISMISSED"
            ]
          },
          "proposalId": {
            "type": "object",
            "nullable": true,
            "description": "Proposal built from this detection, if any"
          },
          "xeroInvoiceId": {
            "type": "object",
            "nullable": true,
            "description": "Xero invoice id once invoiced"
          },
          "createdAt": {
            "type": "string",
            "description": "ISO created timestamp"
          }
        },
        "required": [
          "id",
          "type",
          "clientName",
          "valueGbp",
          "summary",
          "source",
          "evidence",
          "state",
          "createdAt"
        ]
      },
...
      "DetectionProposeResultDto": {
        "type": "object",
        "properties": {
          "detectionId": {
            "type": "string"
          },
          "detectionState": {
            "type": "string",
            "enum": [
              "OPEN",
              "PROPOSED",
              "RESOLVED",
              "DISMISSED"
            ]
          },
          "proposalId": {
            "type": "string",
            "description": "The invoice proposal built from this detection"
          },
          "proposalReference": {
            "type": "object",
            "nullable": true,
            "description": "Idempotency reference used in Xero"
          },
          "subtotal": {
            "type": "number",
            "example": 4200
          },
          "taxTotal": {
            "type": "number",
            "example": 840
          },
          "total": {
            "type": "number",
            "example": 5040
          },
          "taskId": {
            "type": "object",
            "nullable": true,
            "description": "The REVIEW_INVOICE task raised for the human"
          },
          "message": {
            "type": "string",
            "description": "Human sentence describing what happened"
          }
        },
        "required": [
          "detectionId",
          "detectionState",
          "proposalId",
          "subtotal",
          "taxTotal",
          "total",
          "message"
        ]
      },
      "DismissDetectionDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "example": "Client was billed for this separately in cash.",
            "description": "Why the detection was dismissed (kept on the audit trail)"
          }
        }
      },
      "DetectionActionResultDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "state": {
            "type": "string",
            "enum": [
              "OPEN",
              "PROPOSED",
              "RESOLVED",
              "DISMISSED"
            ]
          },
          "message": {
            "type": "string",
            "description": "Human sentence describing what happened"
          }
        },
        "required": [
          "id",
          "state",
          "message"
        ]
      },
...
    }
  }
}